chore(tests): limit multiregion ci to multiregion tests
