+++ packages/server/.env-example
@@ -10,6 +10,9 @@ PORT=3000
 CANONICAL_URL="http://127.0.0.1:3000"
 SESSION_SECRET="-> FILL IN <-"
 
+# Optional license token for paid features like multiregion
+# LICENSE_TOKEN=
+
 # Redis connection: default for local development environment
 REDIS_URL="redis://127.0.0.1:6379"
