+++ packages/server/modules/multiregion/tests/e2e/serverAdmin.graph.spec.ts
@@ -28,7 +28,7 @@ import { expect } from 'chai'
 const isEnabled = isMultiRegionEnabled()
 
 isEnabled
-  ? describe('Multi Region Server Settings', () => {
+  ? describe('Multi Region Server Settings @multiregion', () => {
       let testAdminUser: BasicTestUser
       let testBasicUser: BasicTestUser
       let apollo: TestApolloServer
