+++ packages/server/modules/workspaces/tests/integration/subs.graph.spec.ts
@@ -93,7 +93,7 @@ describe('Workspace GQL Subscriptions', () => {
   ]
 
   modes.forEach(({ isMultiRegion }) => {
-    describe(`W/${!isMultiRegion ? 'o' : ''} multiregion`, () => {
+    describe(`W/${!isMultiRegion ? 'o' : ''} @multiregion`, () => {
       const myMainWorkspace: BasicTestWorkspace = {
         id: '',
         ownerId: '',
